+++ tailwind.config.ts
@@ -26,6 +26,7 @@ const config: Config = {
           'menu-background': '#202123',
           'menu-text': '#E3E3E8',
           'menu-hover':'#343540',
+          'menu-dropdown-background': '#202123',
           'message-background-dark': '#343540',
           'message-background-light': '#f9fafb', /* gray-50 */
         },