+++ src/components/chat/Menu/FooterMenuContent/ModalPlan.tsx
@@ -1,0 +1,7 @@
+export default function ModalPlan() {
+  return (
+    <>
+      my plan
+    </>
+  )
+}