+++ src/components/chat/Menu/FooterMenuContent/ModalSettings.tsx
@@ -1,0 +1,7 @@
+export default function ModalSettings() {
+  return (
+    <>
+      my settings
+    </>
+  )
+}