+++ src/components/chat/Menu/FooterMenuContent/UserCard.tsx
@@ -4,6 +4,7 @@ interface IProps {
   imageSrc: string
   imageAlt: string
   text: string
+  onClick: () => void
 }
 
 export const userData = [
@@ -12,11 +13,14 @@ export const userData = [
 ]
 
 export default function UserCard(props: IProps) {
-  const {imageSrc, imageAlt, text} = props
+  const {imageSrc, imageAlt, text, onClick} = props
 
   return (
     <>
-      <div className={'flex items-center gap-4 p-3 rounded-md cursor-pointer hover:bg-gray-200 transition-change'}>
+      <div
+        className={'flex items-center gap-4 p-3 rounded-md cursor-pointer hover:bg-gray-200 transition-change'}
+        onClick={onClick}
+      >
         <Image
           width={20}
           height={20}