+++ src/components/chat/Message/FooterContent/index.tsx
@@ -1,4 +1,4 @@
-import { type KeyboardEvent, useRef } from 'react'
+import { type ChangeEvent, type KeyboardEvent, useRef } from 'react'
 
 import TextArea from '@/components/ui/TextArea'
 import SendIcon from '@/components/icons/SendIcon'
@@ -23,7 +23,7 @@ export default function FooterContent() {
     }
   }
 
-  const handlerChange = (e: any) => {
+  const handlerChange = (e: ChangeEvent<HTMLTextAreaElement>) => {
     const textarea = e.target
     textarea.style.height = 'auto'
     textarea.style.height = Math.min(textarea.scrollHeight, maxHeight) + 'px'
@@ -34,7 +34,10 @@ export default function FooterContent() {
       <TextArea
         ref={ref}
         placeholder={'Type a message ...'}
-        className={'lg:w-[640px] -ml-6 resize-none rounded-xl drop-shadow'}
+        className={`
+          lg:w-[640px] -ml-6 resize-none rounded-xl drop-shadow custom-message-light-scrollbar
+          dark:bg-chatpage-message-background-dark 
+        `}
         onChange={handlerChange}
         onKeyDown={handleKeyDown}
       />
@@ -42,7 +45,10 @@ export default function FooterContent() {
       <SendIcon
         width={24}
         height={24}
-        className={'cursor-pointer drop-shadow text-gray-600 hover:opacity-50 transition-change -ml-12'}
+        className={`
+          cursor-pointer drop-shadow text-gray-600 hover:opacity-50 transition-change -ml-12
+          dark:text-gray-50/80 dark:hover:opacity-70
+        `}
         onClick={handleSubmit}
       />
     </div>