+++ src/components/chat/Message/HeaderContent/index.tsx
@@ -15,13 +15,18 @@ export default function HeaderContent() {
   const isOpen = useAtomValue(isOpenAtom)
 
   return (
-    <div className={'w-full h-[66px] p-3 border-b flex justify-between items-center'}>
+    <div
+      className={`
+        w-full h-[66px] p-3 border-b mb-1 flex justify-between items-center 
+        dark:bg-chatpage-message-background-dark dark:border-b-gray-500
+      `}
+    >
       {/* left icon */}
       <div className={`${isOpen ? '' : 'ml-12'}`}>
         <Select
           width={`w-fit`}
           options={options}
-          className={'text-black'}
+          className={'text-black dark:text-gray-50 dark:hover:bg-gray-500/20'}
         />
       </div>