+++ src/components/chat/Message/index.tsx
@@ -4,7 +4,7 @@ import FooterContent from '@/components/chat/Message/FooterContent'
 
 export default function Message() {
   return (
-    <div className={'flex-1 flex flex-col bg-gray-50'}>
+    <div className={'flex-1 flex flex-col bg-gray-50 dark:bg-chatpage-message-background-dark'}>
       <HeaderContent/>
       <MainContent/>
       <FooterContent/>