+++ src/components/icons/LanguageIcon.tsx
@@ -1,0 +1,22 @@
+interface IProps {
+  width: number
+  height: number
+  className?: string
+  onClick?: () => void
+}
+
+export default function LanguageIcon(props: IProps) {
+  const {width, height, className, onClick} = props
+
+  return (
+    <div className={className} onClick={onClick}>
+      <svg xmlns="http://www.w3.org/2000/svg" width={width} height={height} viewBox="0 0 24 24" fill="none"
+           stroke="currentColor" strokeWidth="1" strokeLinecap="round" strokeLinejoin="round"
+           className="feather feather-globe">
+        <circle cx="12" cy="12" r="10"/>
+        <line x1="2" y1="12" x2="22" y2="12"/>
+        <path d="M12 2a15.3 15.3 0 0 1 4 10 15.3 15.3 0 0 1-4 10 15.3 15.3 0 0 1-4-10 15.3 15.3 0 0 1 4-10z"/>
+      </svg>
+    </div>
+  )
+}