+++ src/components/ui/Select/index.tsx
@@ -1,8 +1,9 @@
-import Image from 'next/image'
 import { motion } from 'framer-motion'
 import { ReactNode, useRef, useState } from 'react'
 
 import useOutsideClick from '@/hooks/useOutsideClick'
+import ChevronUpIcon from '@/components/icons/ChevronUpIcon'
+import ChevronDownIcon from '@/components/icons/ChevronDownIcon'
 
 type TOptions = {
   value: ReactNode
@@ -36,14 +37,18 @@ export default function Select({
       ref={wrapperRef}
     >
       <div
-        className={'flex gap-2 p-2 cursor-pointer border border-gray-200 rounded-md hover:bg-gray-200 transition-change'}
+        className={`
+          flex gap-2 p-2 cursor-pointer border border-gray-200
+          rounded-md hover:bg-gray-200 transition-change
+          dark:border-gray-500 dark:hover:bg-gray-500/20
+        `}
         onClick={() => setIsOpen(!isOpen)}
       >
         {selectedOption}
         {isOpen ? (
-          <Image alt={'arrow Up'} width={16} height={16} src={'/chevron-up.svg'}/>
+          <ChevronUpIcon width={16} height={16} className={'flex items-center'}/>
         ): (
-          <Image alt={'arrow down'} width={16} height={16} src={'/chevron-down.svg'}/>
+          <ChevronDownIcon width={16} height={16} className={'flex items-center'}/>
         )}
       </div>
 
@@ -52,13 +57,17 @@ export default function Select({
           initial={{ opacity: 0, y: -10 }}
           animate={{ opacity: 1, y: 0 }}
           exit={{ opacity: 0, y: -10 }}
-          className={'absolute z-50 w-48 bg-gray-50 mt-1 border border-gray-200 rounded-md p-2 transition-change'}
+          className={`
+            absolute z-50 w-48 bg-gray-50 mt-1 
+            border border-gray-200 rounded-md p-2 transition-change
+            dark:bg-chatpage-message-background-dark dark:border-gray-500
+          `}
         >
           {options.map((option, index) => (
             <motion.li
               key={index}
               onClick={() => handleSelect(option)}
-              className={'cursor-pointer rounded-md py-2 hover:bg-gray-200 transition-change'}
+              className={'cursor-pointer rounded-md py-2 hover:bg-gray-200 dark:hover:bg-gray-500/20 transition-change'}
             >
               {option.value}
             </motion.li>
@@ -68,11 +77,3 @@ export default function Select({
     </div>
   )
 }
-
-
-
-
-
-
-
-