+++ src/components/ui/TextArea/index.tsx
@@ -12,9 +12,9 @@ const TextArea = forwardRef<HTMLTextAreaElement, TextareaProps>((
       <textarea
         ref={ref}
         className={`
-          flex min-h-[40px] w-full rounded-md border border-input bg-background px-3 py-2 text-sm
+          flex min-h-[40px] w-full rounded-md border border-input px-3 py-2 text-sm
           ring-offset-background placeholder:text-muted-foreground
-          focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2
+          focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring focus-visible:ring-offset-1
           disabled:cursor-not-allowed disabled:opacity-50
           ${className}
         `}