+++ src/types/chat.ts
@@ -1,0 +1,8 @@
+export type TChatItem = {
+  id: number
+  itemName: string
+  itemUuid: string
+  createdAt: string
+  updatedAt: string
+}
+