+++ src/types/index.ts
@@ -1,8 +1,1 @@
-export type TChatItem = {
-  id: number
-  itemName: string
-  itemUuid: string
-  createdAt: string
-  updatedAt: string
-}
-
+export * from './chat'